apps/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
geoalchemy2
pydantic-settings
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
geoalchemy2
pydantic-settings
redis
shapely
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
email-validator
aiosqlite
python-multipart
websockets